Backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
pinecone-client
openai
pandas
numpy
reportlab
authlib
